requirements.txt
pandas==2.0.3
SQLAlchemy==2.0.20
psycopg2-binary==2.9.6
boto3==1.26.50

